refactor(xapiri): extract tab_config, tab_provision, finish tab_editor
Mechanical move per ADR 0014. No behaviour change.
- tab_config.go: updateConfigTab, updateCfgListScreen, updateCfgNewNameScreen,
  renderConfigTab, renderCfgListScreen, renderCfgNewNameScreen
- tab_provision.go: updateProvisionTab, updateCfgEditScreen,
  renderProvisionTab, renderCfgEditScreen
- tab_editor.go: add resolveEditor, renderEditorPlaceholder (missed in prior commit)
