+++ internal/ui/xapiri/dashboard.go
@@ -1070,159 +1070,7 @@ func (m dashModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	return m, nil
 }
 
-// updateConfigTab dispatches key events to the active config sub-screen.
-func (m dashModel) updateConfigTab(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch m.cfgScreen {
-	case cfgScreenList:
-		return m.updateCfgListScreen(msg)
-	case cfgScreenNewName:
-		return m.updateCfgNewNameScreen(msg)
-	}
-	return m, nil
-}
 
-// updateProvisionTab handles key events on the provision tab (full edit form).
-func (m dashModel) updateProvisionTab(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	return m.updateCfgEditScreen(msg)
-}
-
-// updateCfgListScreen handles keys on the config-list screen.
-func (m dashModel) updateCfgListScreen(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	if m.cfgLoading {
-		return m, nil
-	}
-	key := msg.Type
-	keyStr := msg.String()
-	total := len(m.cfgCandidates) + 1 // +1 for the "[ + New config ]" sentinel
-	switch {
-	case key == tea.KeyUp:
-		if m.cfgListCursor > 0 {
-			m.cfgListCursor--
-		}
-	case key == tea.KeyDown:
-		if m.cfgListCursor < total-1 {
-			m.cfgListCursor++
-		}
-	case key == tea.KeyEnter:
-		if m.cfgListCursor == len(m.cfgCandidates) {
-			// New config sentinel selected.
-			m.cfgScreen = cfgScreenNewName
-			m.cfgNewInput.SetValue("")
-			cmd := m.cfgNewInput.Focus()
-			return m, cmd
-		}
-		c := m.cfgCandidates[m.cfgListCursor]
-		m.cfgLoading = true
-		m.cfgLoadErr = ""
-		return m, m.loadCfgEntryCmd(c)
-	case keyStr == "n":
-		m.cfgScreen = cfgScreenNewName
-		m.cfgNewInput.SetValue("")
-		cmd := m.cfgNewInput.Focus()
-		return m, cmd
-	case keyStr == "r":
-		m.cfgLoading = true
-		m.cfgLoadErr = ""
-		m.cfgCandidates = nil
-		return m, m.loadCfgListCmd()
-	}
-	return m, nil
-}
-
-// updateCfgNewNameScreen handles keys while entering a new config name.
-func (m dashModel) updateCfgNewNameScreen(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	key := msg.Type
-	switch {
-	case key == tea.KeyEsc:
-		m.cfgScreen = cfgScreenList
-		m.cfgNewInput.Blur()
-		return m, nil
-	case key == tea.KeyEnter:
-		name := strings.TrimSpace(m.cfgNewInput.Value())
-		m.cfgNewInput.Blur()
-		if name != "" {
-			m.cfg.ConfigName = name
-			m.cfg.ConfigNameExplicit = true
-		}
-		m.cfgSelected = true
-		m.activeTab = tabProvision
-		return m, textinput.Blink
-	default:
-		var cmd tea.Cmd
-		m.cfgNewInput, cmd = m.cfgNewInput.Update(msg)
-		return m, cmd
-	}
-}
-
-// updateCfgEditScreen handles key events in the full config edit form.
-func (m dashModel) updateCfgEditScreen(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	// Discard keystrokes while a config entry is loading to prevent input
-	// from being silently lost when cfgEntryLoadMsg rebuilds all inputs.
-	if m.cfgLoading {
-		return m, nil
-	}
-	key := msg.Type
-	keyStr := msg.String()
-
-	switch {
-	case keyStr == "ctrl+l":
-		m.activeTab = tabLogs
-		return m, nil
-
-	case key == tea.KeyUp:
-		m = m.moveFocus(false)
-		return m, textinput.Blink
-
-	case key == tea.KeyDown:
-		m = m.moveFocus(true)
-		return m, textinput.Blink
-	}
-
-	// Per-field-kind handling.
-	meta := dashFields[m.focus]
-	switch meta.kind {
-	case fkToggle:
-		if key == tea.KeySpace || key == tea.KeyEnter {
-			m.toggles[meta.subIdx] = !m.toggles[meta.subIdx]
-			if meta.costKey {
-				return m, m.markDirty()
-			}
-		}
-
-	case fkSelect:
-		switch {
-		case key == tea.KeyRight || key == tea.KeyEnter || keyStr == "l":
-			m.selects[meta.subIdx].next()
-			if meta.subIdx == siMode {
-				m = m.rebuildProviderList()
-			}
-			if meta.costKey {
-				return m, m.markDirty()
-			}
-		case key == tea.KeyLeft || keyStr == "h":
-			m.selects[meta.subIdx].prev()
-			if meta.subIdx == siMode {
-				m = m.rebuildProviderList()
-			}
-			if meta.costKey {
-				return m, m.markDirty()
-			}
-		}
-
-	case fkText:
-		if key == tea.KeyTab || key == tea.KeyShiftTab {
-			return m, nil
-		}
-		ti, cmd := m.textInputs[meta.subIdx].Update(msg)
-		m.textInputs[meta.subIdx] = ti
-		if meta.costKey {
-			return m, tea.Batch(cmd, m.markDirty())
-		}
-		return m, cmd
-	}
-
-	return m, nil
-}
 
 
 
@@ -1273,170 +1121,6 @@ func (m dashModel) View() string {
 	}
 
 	return lipgloss.JoinVertical(lipgloss.Left, tabBar, content, termPane, bottomStrip, footer)
-}
-
-// ─── config tab ──────────────────────────────────────────────────────────────
-
-
-// renderConfigTab dispatches to the active config sub-screen renderer.
-func (m dashModel) renderConfigTab(w, h int) string {
-	switch m.cfgScreen {
-	case cfgScreenNewName:
-		return m.renderCfgNewNameScreen(w, h)
-	default:
-		return m.renderCfgListScreen(w, h)
-	}
-}
-
-// renderProvisionTab renders the full interactive provision edit form.
-func (m dashModel) renderProvisionTab(w, h int) string {
-	return m.renderCfgEditScreen(w, h)
-}
-
-// renderCfgListScreen renders the list of existing bootstrap configs.
-func (m dashModel) renderCfgListScreen(w, h int) string {
-	var lines []string
-	title := fmt.Sprintf(" Configurations on kind cluster %q", m.cfg.KindClusterName)
-	lines = append(lines, stHdr.Render(title))
-	lines = append(lines, stMuted.Render(strings.Repeat("─", w)))
-	lines = append(lines, "")
-
-	if m.cfgLoading {
-		lines = append(lines, stMuted.Render("  loading…"))
-	} else if m.cfgLoadErr != "" {
-		lines = append(lines, stBad.Render("  "+m.cfgLoadErr))
-		lines = append(lines, stMuted.Render("  r = retry"))
-	} else {
-		for i, c := range m.cfgCandidates {
-			cursor := "  "
-			if i == m.cfgListCursor {
-				cursor = stAccent.Render("▸ ")
-			}
-			lines = append(lines, cursor+c.Label())
-		}
-		lines = append(lines, "")
-		// "[ + New config ]" sentinel.
-		sentinelIdx := len(m.cfgCandidates)
-		newLabel := "[ + New config ]"
-		if m.cfgListCursor == sentinelIdx {
-			lines = append(lines, stAccent.Render("▸ ")+stAccent.Render(newLabel))
-		} else {
-			lines = append(lines, "  "+stMuted.Render(newLabel))
-		}
-	}
-
-	lines = append(lines, "")
-	if m.cfgLoading {
-		lines = append(lines, stMuted.Render("  please wait…"))
-	} else {
-		lines = append(lines, stMuted.Render("  ↑/↓  navigate    enter  select    n  new config    r  refresh    esc/q  quit"))
-	}
-	lines = append(lines, "")
-	if m.cfgSelected {
-		lines = append(lines, stOK.Render("  config: "+m.cfg.ConfigName+"  — select another to switch, or press 2 to return to provision"))
-	} else {
-		lines = append(lines, stWarn.Render("  ⚠ Select a configuration to unlock all tabs"))
-	}
-
-	for len(lines) < h {
-		lines = append(lines, "")
-	}
-	return strings.Join(lines[:min(len(lines), h)], "\n")
-}
-
-// renderCfgNewNameScreen renders the new config name input.
-func (m dashModel) renderCfgNewNameScreen(w, h int) string {
-	var lines []string
-	lines = append(lines, stHdr.Render(" New configuration name"))
-	lines = append(lines, stMuted.Render(strings.Repeat("─", w)))
-	lines = append(lines, "")
-	lines = append(lines, stMuted.Render("  Enter a name for the new configuration."))
-	lines = append(lines, stMuted.Render("  A dedicated namespace yage-<name> will be created on the kind cluster."))
-	lines = append(lines, "  (leave blank to use the workload cluster name as the default)")
-	lines = append(lines, "")
-	lines = append(lines, "  "+m.cfgNewInput.View())
-	lines = append(lines, "")
-	lines = append(lines, stMuted.Render("  enter  confirm    esc  back to list"))
-
-	for len(lines) < h {
-		lines = append(lines, "")
-	}
-	return strings.Join(lines[:min(len(lines), h)], "\n")
-}
-
-// renderCfgEditScreen renders the full interactive config edit form.
-func (m dashModel) renderCfgEditScreen(w, h int) string {
-	var lines []string
-	lastSection := ""
-
-	// Config name badge at the top.
-	if name := m.cfg.ConfigName; name != "" {
-		lines = append(lines, stMuted.Render(fmt.Sprintf("  config: %s", name)))
-		lines = append(lines, "")
-	}
-
-	for fid := 0; fid < focCount; fid++ {
-		if m.isHidden(fid) {
-			continue
-		}
-		meta := dashFields[fid]
-		// Section header.
-		if meta.section != "" && meta.section != lastSection {
-			if len(lines) > 0 {
-				lines = append(lines, "")
-			}
-			lines = append(lines, stHdr.Render(" "+meta.section))
-			lastSection = meta.section
-		}
-
-		focused := m.focus == fid
-		lines = append(lines, m.renderField(fid, focused, w))
-	}
-
-	// Pad to height.
-	for len(lines) < h {
-		lines = append(lines, "")
-	}
-
-	return strings.Join(lines[:min(len(lines), h)], "\n")
-}
-
-// resolveEditor returns the editor binary to use for in-place editing.
-// Priority: $VISUAL → $EDITOR → first hit in editorFallbacks (OS-specific).
-// Env-var values and fallback candidates are probed with exec.LookPath
-// where possible. If none are found, it returns a conventional fallback
-// name for exec to report on, so it never returns an empty string.
-func resolveEditor() string {
-	for _, env := range []string{"VISUAL", "EDITOR"} {
-		if v := strings.TrimSpace(os.Getenv(env)); v != "" {
-			if p, err := exec.LookPath(v); err == nil {
-				return p
-			}
-			// Env var set but binary not in PATH — fall through to probe list.
-		}
-	}
-	for _, candidate := range editorFallbacks {
-		if p, err := exec.LookPath(candidate); err == nil {
-			return p
-		}
-	}
-	// Last resort: return the final fallback name even if not found —
-	// exec will produce a clear error message to the user.
-	if len(editorFallbacks) > 0 {
-		return editorFallbacks[len(editorFallbacks)-1]
-	}
-	return "vi"
-}
-
-// renderEditorPlaceholder shows while waiting for the editor to launch.
-func (m dashModel) renderEditorPlaceholder(w, h int) string {
-	editor := resolveEditor()
-	msg := stMuted.Render(fmt.Sprintf("  Opening %s…  (press any key after it exits)", editor))
-	lines := []string{"", msg}
-	for len(lines) < h {
-		lines = append(lines, "")
-	}
-	return strings.Join(lines, "\n")
 }
